test: exercise real directive code in landing page test
Import production directive definitions (vReveal, vCount) instead of
no-op stubs so the test exercises the real code path.  The directive
logic is extracted into utils/directives.ts so it can be shared between
the Nuxt plugin and tests without pulling in defineNuxtPlugin.
